aws설정시간

+++ src/main/java/org/zerock/service/board/BoardService.java
@@ -169,11 +169,13 @@ public class BoardService {
 		
 		File[] listFiles = folder.listFiles();
 		
-		for (File file : listFiles) {
-			file.delete();
+		if(listFiles != null) {		
+			for (File file : listFiles) {
+				file.delete();
+			}
+			
+			folder.delete();
 		}
-		
-		folder.delete();
 		
 		// db 파일 records 지우기
 		boardMapper.deleteFileByBoardId(id);
